moved categories to directory component and added category routing
